Move encoding utils to separate file
